% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{bayes.PHMM}
\alias{bayes.PHMM}
\title{Poisson Hidden Markov Model}
\usage{
bayes.PHMM(y, m = 2, ...)
}
\arguments{
\item{y}{observed series.}

\item{m}{the number of states.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
Fit a Bayesian Poisson hidden Markov model with stan.
}
\examples{

data("homicides") # load data
Homicides <- ts(data = round(homicides$Rate), start = 1960) # convert to time series

BayesPHHMM_2states <- bayes.PHMM(y = Homicides, m = 2, chains = 1, iter = 500,
                                control = list(adapt_delta = 0.99), seed = 4)

}
