% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{bayes.ZIPHMM}
\alias{bayes.ZIPHMM}
\title{ZIP-HMM (Zero inflated all States)}
\usage{
bayes.ZIPHMM(y, m = 2, ...)
}
\arguments{
\item{y}{observed series.}

\item{m}{the number of states.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
Fit a Bayesian Zero inflated Poisson hidden Markov model with stan.
}
