% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bayeshmmcts-package.R
\docType{package}
\name{Bayeshmmcts-package}
\alias{Bayeshmmcts-package}
\alias{Bayeshmmcts}
\title{Bayeshmmcts: Fit hidden Markov Models via RStan.}
\description{
Fast and flexible tool for Bayesian estimation of hidden Markov models using the 'rstan' package, which provides the R interface to the Stan C++ library.
}
\references{
Stan Development Team (2020). RStan: the R interface to Stan. R package version 2.21.2. https://mc-stan.org
}
