% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{homicides}
\alias{homicides}
\title{Data from Homicidios (murder and murder rate)}
\format{
A data frame with 50 observations and 4 variables:
\describe{
  \item{Year}{Year in which the homicide was reported.}
  \item{Murder}{Total number of homicides registered for a given year.}
  \item{Population}{Total population in Colombia for a given year.}
  \item{Rate}{Homicide rate per year per 100.000 inhabitants.}
  ...
}
}
\usage{
homicides
}
\description{
This data contains the total number of homicides and the intentional homicide
rate in Colombia for the years 1960 to 2018.
The data were drawn from the National Planning Department (DNP) and the
criminal statistics of the National Police.
}
\keyword{datasets}
