% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.conditional}
\alias{pois.HMM.conditional}
\title{Conditional probabilities}
\usage{
pois.HMM.conditional(oc, o, mod)
}
\arguments{
\item{oc}{A input specifies the range of o-values for which these probabilities are required.}

\item{o}{observed series.}

\item{mod}{a Poisson hidden Markov model.}
}
\value{
the result is a matrix with conditional probabilities.
}
\description{
Conditional probability that observation at time \eqn{t} equals \code{oc},
given all observations other than that at time \eqn{t}.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
pois.HMM.conditional(oc = 10:15, o = o, mod = mod_2states)

}
