% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.forecast}
\alias{pois.HMM.forecast}
\title{Forecast probabilities}
\usage{
pois.HMM.forecast(of, h = 1, o, mod)
}
\arguments{
\item{of}{The range of \code{o}-values for which these probabilities are required is specified.}

\item{h}{The range of times for which they are required.}

\item{o}{observed series.}

\item{mod}{a Poisson hidden Markov model.}
}
\value{
output is a matrix with the forecast probabilities.
}
\description{
Computing forecast probabilities using a PHMM.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
forecasts <- pois.HMM.forecast(of = 0:50, h = 1, o, mod_2states)
par(mfrow = c(1, 1), las = 1)
plot(0:50, forecasts, type = "h", main = paste("forecast distribution for", length(o) + 1),
xlim = c(0, max(0:50)), ylim = c(0, 0.12), xlab = "count", ylab = "probability", lwd = 3)

}
