% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.generate_sample}
\alias{pois.HMM.generate_sample}
\title{Generating a sample}
\usage{
pois.HMM.generate_sample(ns, mod)
}
\arguments{
\item{ns}{size of sample.}

\item{mod}{a Poisson hidden Markov model, from \code{pois.HMM.mle} function.}
}
\value{
x a random values from Poisson hidden Markov model.
}
\description{
This function generates a realization, of length ns, of the PHMM mod.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
pois.HMM.generate_sample(ns = 10, mod = mod_2states)

}
