% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.lforward}
\alias{pois.HMM.lforward}
\title{Computing log(forward probabilities)}
\usage{
pois.HMM.lforward(o, mod)
}
\arguments{
\item{o}{observed series.}

\item{mod}{a Poisson hidden Markov model.}
}
\value{
A matrix (lalpha) is returned.
}
\description{
Given data o and model mod, this function uses the recursion \eqn{\alpha_{t+1} = \alpha_t A P(o_{t+1})} to find all the vectors of forward probabilities, in logarithmic form.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
pois.HMM.lforward(o = o, mod = mod_2states)

}
