% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.local_decoding}
\alias{pois.HMM.local_decoding}
\title{Local decoding}
\usage{
pois.HMM.local_decoding(o, mod)
}
\arguments{
\item{o}{observed series.}

\item{mod}{a Poisson hidden Markov model.}
}
\value{
A vector with the most likely states.
}
\description{
This approach determines the most likely sequence of (hidden) state separately for each \eqn{t}.
}
\details{
The local decoding in this approach consists in determines the most likely state separately for each \eqn{t} by maximizing the conditional probability.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
pois.HMM.local_decoding(o = o, mod = mod_2states)

}
