% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.mle}
\alias{pois.HMM.mle}
\title{Computing the MLEs, given starting values for the natural parameters}
\usage{
pois.HMM.mle(o, m, lambda0, A0, pi0 = NULL, stationary = TRUE, ...)
}
\arguments{
\item{o}{a non-negative vector with of count data time serie.}

\item{m}{the number of states.}

\item{lambda0}{starting value of \eqn{\lambda}, the vector of state-dependent means in a Poisson HMM.}

\item{A0}{starting value of \eqn{A}, the transition probability matrix.}

\item{pi0}{starting value of \eqn{\pi}, the stationary distribution of the Markov chain.}

\item{stationary}{indicate if the Markov chain is stationary, by default is \code{TRUE}.}

\item{...}{Arguments passed to `stats::nlm` (e.g. gradtol, steptol, iterlim).}
}
\value{
A list of parameters
\itemize{
  \item m - the number of states.
  \item lambda - estimates values of \eqn{\lambda}.
  \item A - estimates values of \eqn{A}.
  \item pi - estimates values of \eqn{\pi}.
  \item code - an integer indicating why the optimization process terminated. For more details see \code{nlm{stats}}.
  \item mllk - Minus the log-likelihood.
  \item AIC - The Akaike information criterion (AIC).
  \item BIC - the Bayesian information criterion (BIC)
}
}
\description{
Estimate the (natural) parameters of the model by using numerical minimization
of minus the log-likelihood. For more details see the function `nlm{stats}`.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
acf(o)
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
print(mod_2states)

}
