% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.mllk}
\alias{pois.HMM.mllk}
\title{Computing minus the log-likelihood from the working parameters}
\usage{
pois.HMM.mllk(parvect, o, m, stationary = TRUE)
}
\arguments{
\item{parvect}{parameters vector of working from \code{pois.HMM.pn2pw} function.}

\item{o}{a non-negative vector with of count data time serie.}

\item{m}{the number of states.}

\item{stationary}{indicate if the Markov chain is stationary, by default is \code{TRUE}.}
}
\value{
mllk Minus the log-likelihood.
}
\description{
Compute minus the log-likelihood of a Poisson HMM for given values of the working parameters.
}
\examples{

m0 = 2
lambda0 = c(15, 25)
A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE)
pi0 = c(0.5,0.5)
parvect_work <- pois.HMM.pn2pw(m = m0, lambda = lambda0, A = A0, pi = pi0, stationary = TRUE)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
pois.HMM.mllk(parvect = parvect_work, o = o, m = 2, stationary = TRUE)

}
