% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.moments}
\alias{pois.HMM.moments}
\title{Poisson HMM moments}
\usage{
pois.HMM.moments(mod, lag.max = 10)
}
\arguments{
\item{mod}{a Poisson hidden Markov model.}

\item{lag.max}{Maximum lag for correlation function \eqn{\rho}.}
}
\value{
A list of moments and rho
\itemize{
  \item mu - estimate value of first moment \eqn{\mu}.
  \item sigma - estimate value of variance \eqn{\sigma}.
  \item rho - estimates values of \eqn{\rho}, for a lag \eqn{h}.
}
}
\description{
Computing moment from a Poisson hidden Markov model, with stationary Markov chain.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
pois.HMM.moments(mod = mod_2states)

}
