% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.plot.residuals}
\alias{pois.HMM.plot.residuals}
\title{Plotting pseudo-residuals}
\usage{
pois.HMM.plot.residuals(residual, ...)
}
\arguments{
\item{residual}{is a ordinary normal pseudo-residuals.}

\item{...}{further arguments and graphical parameters passed to plot.histogram and thence to title and axis (if plot = TRUE).}
}
\value{
four plots (index, qq plot, histogram and ACF).
}
\description{
Plot of the ordinary normal mid-pseudo-residuals
}
\details{
The top left shows index plot of the normal pseudo-residuals, with horizontal lines at 0, \eqn{\pm 1.96}, \eqn{\pm 2.58}. The top right shows quantile-quantile plots of the normal pseudo-residuals, with the theoretical quantiles on the horizontal axis. The bottom left show histograms of the normal pseudo-residuals. The last plot shows the autocorrelation functions of the normal pseudo-residuals.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
res <- pois.HMM.pseudo_residuals(o = o, mod = mod_2states)
pois.HMM.plot.residuals(res)

}
