% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.pn2pw}
\alias{pois.HMM.pn2pw}
\title{Transforming natural parameters to working}
\usage{
pois.HMM.pn2pw(m, lambda, A, pi = NULL, stationary = TRUE)
}
\arguments{
\item{m}{the number of states.}

\item{lambda}{the estimates of \eqn{\lambda}.}

\item{A}{the estimates of \eqn{A}.}

\item{pi}{the estimates of \eqn{\pi}.}

\item{stationary}{indicate if the Markov chain is stationary, by default is \code{TRUE}.}
}
\value{
parvect parameters vector of working.
}
\description{
The purpose of the transformation of the natural parameters (Poisson means, transition probabilities and, if appropriate, the initial distribution of the Markov chain) to working parameters is simply to convert a constrained optimization problem to an unconstrained one.
}
\examples{

m0 = 2
lambda0 = c(15, 25)
A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE)
pi0 = c(0.5,0.5)
parvect_work <- pois.HMM.pn2pw(m = m0, lambda = lambda0, A = A0, pi = pi0, stationary = T)
print(parvect_work)

}
