% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.pw2pn}
\alias{pois.HMM.pw2pn}
\title{Transforming working parameters to natural}
\usage{
pois.HMM.pw2pn(m, parvect, stationary = TRUE)
}
\arguments{
\item{m}{the number of states.}

\item{parvect}{parameters vector of working from \code{pois.HMM.pn2pw} function.}

\item{stationary}{indicate if the Markov chain is stationary, by default is \code{TRUE}.}
}
\value{
A list of parameters
\itemize{
  \item lambda - estimates values of \eqn{\lambda}.
  \item A - estimates values of \eqn{A}.
  \item pi - estimates values of \eqn{\pi}.
}
}
\description{
Transform the estimates of the working parameters to estimates of
the natural parameters.
}
\examples{

m0 = 2
lambda0 = c(15, 25)
A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE)
pi0 = c(0.5,0.5)
parvect_work <- pois.HMM.pn2pw(m = m0, lambda = lambda0, A = A0, pi = pi0, stationary = TRUE)
pois.HMM.pw2pn(m = 2, parvect = parvect_work)

}
