% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.state_prediction}
\alias{pois.HMM.state_prediction}
\title{State prediction}
\usage{
pois.HMM.state_prediction(h = 1, o, mod)
}
\arguments{
\item{h}{Number of periods for forecasting.}

\item{o}{observed series.}

\item{mod}{a Poisson hidden Markov model.}
}
\value{
The state output is a matrix even if h=1.
}
\description{
State prediction using a PHMM.
}
\details{
Compute probabilities using equation \eqn{Pr(C_{T+h} = i | O^{(T)} = o^{(T)})}, for a range of values \eqn{h \in \mathbb{N}}.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
pois.HMM.state_prediction(h = 5, o = o, mod = mod_2states)

}
