% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{pois.HMM.state_probs}
\alias{pois.HMM.state_probs}
\title{State probabilities}
\usage{
pois.HMM.state_probs(o, mod, digits = 7)
}
\arguments{
\item{o}{observed series.}

\item{mod}{a Poisson hidden Markov model.}

\item{digits}{integer indicating the number of decimal places (round).}
}
\value{
A matrix with state probabilities.
}
\description{
Computing state probabilities using log(forward) and log(backward) probabilities.
}
\details{
Here we compute probabilities \eqn{Pr(C_t = i | O^{(T)} = o^{(T)})}, for \eqn{t \in \lbrace 1,2,...,T \rbrace}.
}
\examples{

set.seed(2019)
o <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25))
mod_2states <- pois.HMM.mle(o = o, m = 2, lambda0 = c(20 ,25),
                            A0 = matrix(c(0.9, 0.1, 0.1, 0.9), 2, 2, byrow = TRUE),
                            pi0 = c(0.5,0.5), stationary = TRUE)
pois.HMM.state_probs(o = o, mod = mod_2states)

}
