% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wildfires}
\alias{wildfires}
\title{Data from Incendios Forestales (wildfires)}
\format{
A data frame with 180 rows and 2 variables:
\describe{
  \item{Date}{Year and month in which the wildfire was reported.}
  \item{GIF}{Grandes Incendios Forestales. Large wildfires are defined as those fires that exceed 500 acres affected forest.}
  ...
}
}
\usage{
wildfires
}
\description{
These data describe the number of large forest fires that occurred in Colombia,
from January 2002 to December 2016. The data was extracted from the Institute
of Hydrology, Meteorology and Environmental Studies (IDEAM), and the periodicity
of the data is monthly.
}
\keyword{datasets}
