% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BayesHMM_F.R
\name{bayes.PHMM}
\alias{bayes.PHMM}
\title{Poisson Hidden Markov Model}
\usage{
bayes.PHMM(y, m = 2, ...)
}
\arguments{
\item{y}{observed series.}

\item{m}{the number of states.}

\item{...}{Arguments passed to \code{rstan::sampling} (e.g. iter, chains).}
}
\value{
An object of class \code{stanfit} returned by \code{rstan::sampling}.
}
\description{
Fit a Bayesian Poisson hidden Markov model with stan.
}
\examples{

set.seed(2019)
y <- c(rpois(n = 100, lambda = 15),rpois(n = 50, lambda = 25),rpois(n = 75, lambda = 32))
PHHMM_3states <- bayes.PHMM(y = y, m = 3, chains = 2, iter = 1000)
print(PHHMM_3states, digits_summary = 3)

}
